# Add bootstrap check for JVM tools