--- a/elastic-apm-agent/src/main/java/co/elastic/apm/agent/premain/BootstrapChecks.java
+++ b/elastic-apm-agent/src/main/java/co/elastic/apm/agent/premain/BootstrapChecks.java
@@ -36,7 +36,9 @@ class BootstrapChecks {
 
     static BootstrapChecks defaults() {
         return new BootstrapChecks(!Boolean.parseBoolean(System.getProperty("elastic.apm.disable_bootstrap_checks")),
-            new JavaVersionBootstrapCheck(JvmRuntimeInfo.ofCurrentVM()), new VerifyNoneBootstrapCheck(ManagementFactory.getRuntimeMXBean()));
+            new JavaVersionBootstrapCheck(JvmRuntimeInfo.ofCurrentVM()),
+            new VerifyNoneBootstrapCheck(ManagementFactory.getRuntimeMXBean()),
+            new JvmToolBootstrapCheck(System.getProperty("sun.java.command")));
     }
 
     /**
@@ -57,7 +59,7 @@ class BootstrapChecks {
             if (bootstrapChecksEnabled) {
                 isPassing = false;
                 System.err.println("[elastic-apm-agent] WARN Failed to start agent because of failing bootstrap checks.");
-                System.err.println("[elastic-apm-agent] INFO To override Java version verification, set the 'elastic.apm.disable_bootstrap_checks' System property to 'true'.");
+                System.err.println("[elastic-apm-agent] INFO To override Java bootstrap checks, set the 'elastic.apm.disable_bootstrap_checks' System property to 'true'.");
             } else {
                 System.err.println("[elastic-apm-agent] WARN Bootstrap checks have failed. The agent will still start because bootstrap check have been disabled.");
             }