--- a/elastic-apm-agent/src/main/java/co/elastic/apm/agent/premain/JvmToolBootstrapCheck.java
+++ b/elastic-apm-agent/src/main/java/co/elastic/apm/agent/premain/JvmToolBootstrapCheck.java
@@ -59,7 +59,7 @@ public class JvmToolBootstrapCheck implements BootstrapCheck {
             result.addWarn("Unexpected JVM command line syntax: " + jvmCmd);
         }
         if (isJdkTool) {
-            result.addWarn(String.format("JVM tool detected: '%s' agent will self-disable", jvmCmd));
+            result.addWarn(String.format("JVM tool detected: '%s', agent instrumentation on JVM tools adds unnecessary performance overhead", jvmCmd));
         }
     }
 }