--- a/elastic-apm-agent/src/test/java/co/elastic/apm/agent/premain/JvmToolBootstrapCheckTest.java
+++ b/elastic-apm-agent/src/test/java/co/elastic/apm/agent/premain/JvmToolBootstrapCheckTest.java
@@ -96,15 +96,15 @@ class JvmToolBootstrapCheckTest {
     void testJdkTool(String cmd) {
         BootstrapCheck.BootstrapCheckResult result = new BootstrapCheck.BootstrapCheckResult();
         JvmToolBootstrapCheck.checkJdkTool(cmd, result);
-        assertThat(result.hasErrors())
-            .describedAs("command '%s' should be detected as a JDK tool")
+        assertThat(result.hasWarnings())
+            .describedAs("command '%s' should be detected as a JDK tool", cmd)
                 .isTrue();
 
-        assertThat(result.getErrors()).hasSize(1);
-        assertThat(result.getErrors().get(0)).contains("JVM tool detected");
+        assertThat(result.getWarnings()).hasSize(1);
+        assertThat(result.getWarnings().get(0)).contains("JVM tool detected");
 
-        assertThat(result.hasWarnings())
-            .describedAs("no warning expected")
+        assertThat(result.hasErrors())
+            .describedAs("no error expected")
             .isFalse();
     }
 
@@ -125,22 +125,29 @@ class JvmToolBootstrapCheckTest {
     void checkOtherNotJdkTool(String cmd) {
         BootstrapCheck.BootstrapCheckResult result = new BootstrapCheck.BootstrapCheckResult();
         JvmToolBootstrapCheck.checkJdkTool(cmd, result);
-        assertThat(result.hasErrors())
-            .describedAs("command '%s' should issue an error to help us enhance the heuristic")
+
+        assertThat(result.hasWarnings())
+            .describedAs("command '%s' should issue a warning to help us enhance the heuristic")
             .isTrue();
 
-        assertThat(result.getErrors()).hasSize(1);
-        assertThat(result.getErrors().get(0)).describedAs("warning message should include original command").contains(cmd);
+        assertThat(result.getWarnings()).hasSize(1);
+        assertThat(result.getWarnings().get(0)).describedAs("warning message should include original command").contains(cmd);
+
+        assertThat(result.hasErrors())
+            .describedAs("no error expected")
+            .isFalse();
     }
 
     private static void checkNotJdkTool(@Nullable String cmd) {
         BootstrapCheck.BootstrapCheckResult result = new BootstrapCheck.BootstrapCheckResult();
         JvmToolBootstrapCheck.checkJdkTool(cmd, result);
-        assertThat(result.hasErrors())
+
+        assertThat(result.hasWarnings())
             .describedAs("command '%s' should not be detected as a JDK tool", cmd)
             .isFalse();
-        assertThat(result.hasWarnings())
-            .describedAs("no warning expected")
+
+        assertThat(result.hasErrors())
+            .describedAs("no error expected")
             .isFalse();
     }