## What does this PR do?

In some environments, it might be common to add the agent by using the `JAVA_TOOL_OPTIONS` environment variable to provide the `-javaagent:...` JVM flag.

One downside is that all the JVM tools that also use the JVM will also try to start the agent, hence trying to instrument them and make them slower.

While running `jar` or `javac` or `javadoc` is quite unlikely in production, it is quite common to have a few invocations of `jps` to list JVM processes or `keytool` to import certificates.

When testing this change with all non-interactive JDK 11 binaries in `${JAVA_HOME}/bin` (thus `jconsole`, `jshell`, ... are excluded), we get a significant performance improvement measured through `time` : 

Before:
`71.21s user 2.53s system 249% cpu 29.539 total`

After:
`12.88s user 1.01s system 176% cpu 7.857 total`

For reference, without any agent:
`4.28s user 0.52s system 152% cpu 3.140 total`


## Checklist

- [x] This is a bugfix
  - [x] I have updated [CHANGELOG.asciidoc](https://github.com/elastic/apm-agent-java/blob/main/CHANGELOG.asciidoc)
  - [x] I have added tests that would fail without this fix
- [x] manually testing with JDK 11, 8 and 7 by setting `JAVA_TOOL_OPTIONS` with `-javaagent:...` and try to start all JVM binaries in `${JAVA_HOME}/bin` folder.