@SylvainJuge 2023-01-04
I see what you mean here @eyalkoren: this is just an optimization, thus not something essential that will prevent the agent to work properly (unlike other JVM bootstrap checks). While I like the idea to "fail fast and get quick feedback", that's not really optimal for the end user but more optimized for us.

What I think would be the best in this case would be to:
- make this check issue warnings only, and not errors, thus it won't prevent the agent to start
- ensure that the bootstrap checks results (errors/warnings) are also displayed in `System.out` (or `System.err`) to maximize the chances the end-user will see them if they happen (I think that's already the case but I need to check).

Thus, when a false positive is triggered:
- users are likely to spot the issue if they see it in the standard output, thus they could report to us through support/forum.
- they have a work-around by disabling bootstrap checks

In the case of a false negative, the impact is only visible if the end-user is impacted by a slower-than-usual execution of the JVM tools. Having added tests for most of those tools, I think the probability to have false negatives is low, and would mostly happen through support.